Provide purse identifier to generalize public key in get-account
