+++ lib/cli/parse.rs
@@ -14,8 +14,9 @@ use casper_types::{
 
 use super::{simple_args, CliError, PaymentStrParams, SessionStrParams};
 use crate::{
+    rpcs::common::PurseIdentifier, 
     types::{BlockHash, DeployHash, ExecutableDeployItem, TimeDiff, Timestamp},
-    BlockIdentifier, GlobalStateIdentifier, JsonRpcId, OutputKind, PurseIdentifier, Verbosity,
+    BlockIdentifier, GlobalStateIdentifier, JsonRpcId, OutputKind, Verbosity,
 };
 
 pub(super) fn rpc_id(maybe_rpc_id: &str) -> JsonRpcId {
