+++ lib/rpcs.rs
@@ -6,7 +6,7 @@ pub mod results;
 pub(crate) mod v1_4_5;
 /// RPCs provided by the v1.5.0 node.
 pub(crate) mod v1_5_0;
+pub use common::PurseIdentifier;
 pub use v1_5_0::{
-    get_dictionary_item::DictionaryItemIdentifier, query_balance::PurseIdentifier,
-    query_global_state::GlobalStateIdentifier,
+    get_dictionary_item::DictionaryItemIdentifier, query_global_state::GlobalStateIdentifier,
 };
