+++ lib/rpcs/common.rs
@@ -7,6 +7,7 @@ use crate::types::BlockHash;
 use serde::{Deserialize, Serialize};
 
 use casper_hashing::Digest;
+use casper_types::{account::AccountHash, PublicKey, URef};
 
 /// Enum of possible ways to identify a [`Block`].
 #[derive(Copy, Clone, PartialOrd, Ord, PartialEq, Eq, Hash, Serialize, Deserialize, Debug)]
@@ -31,3 +32,16 @@ pub enum GlobalStateIdentifier {
     /// Query using the state root hash.
     StateRootHash(Digest),
 }
+
+/// Identifier of a purse.
+#[derive(Serialize, Deserialize, Debug, Clone)]
+#[serde(deny_unknown_fields, rename_all = "snake_case")]
+pub enum PurseIdentifier {
+    /// The main purse of the account identified by this public key.
+    MainPurseUnderPublicKey(PublicKey),
+    /// The main purse of the account identified by this account hash.
+    MainPurseUnderAccountHash(AccountHash),
+    /// The purse identified by this URef.
+    PurseUref(URef),
+}
+
