+++ lib/rpcs/v1_4_5/get_account.rs
@@ -4,22 +4,11 @@ use casper_types::{ProtocolVersion, PublicKey};
 
 use crate::{rpcs::common::BlockIdentifier, types::Account};
 
-pub(crate) const GET_ACCOUNT_METHOD: &str = "state_get_account_info";
-
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(deny_unknown_fields)]
 pub(crate) struct GetAccountParams {
     public_key: PublicKey,
     block_identifier: Option<BlockIdentifier>,
-}
-
-impl GetAccountParams {
-    pub(crate) fn new(public_key: PublicKey, block_identifier: Option<BlockIdentifier>) -> Self {
-        GetAccountParams {
-            public_key,
-            block_identifier,
-        }
-    }
 }
 
 /// The `result` field of a successful JSON-RPC response to a `state_get_account_info` request.
